README.md
Hello my name is Kevin, I am Mexican and I am 26 years old. I'm new here because I just decided to take the creative path of software development and when I found out about the existence of GitHub I fell in love with the community interaction that I found in forum posts, videos and articles on the web in general where GitHub stands out as an incubator for startups. It is a pleasure to be here.